requirements.txt
asgiref==3.11.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-rest-auth==7.0.1
Django==4.2.27
django-allauth==65.13.1
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
environ==1.0
google-auth==2.45.0
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==12.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
python-slugify==8.0.4
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.5
text-unidecode==1.3
urllib3==2.6.2
